requirements.txt
streamlit
python-dotenv
google-generativeai>=0.4.0
